refactor(html): use state not href
This commit refactors a ton of client modules to use `ui-sref` tags
which navigate to particular states instead of `href`s which are
dependent on our URL syntax.

Additionally, it deletes two unused files (profit centers and an
accounts template).

It also updates several HTML files to use the translate directive
instead of the translate filter.

Closes #1279.  Closes #1223.