fix(invoices): mv credit note logic into reversed
This commit adds the reversed column which now is the basis for credit
note logic.  This reduces the need for an expensive LEFT JOIN on the
invoices table on every registry load.

There is also a slight performance improvement on the client side as it
is no longer necessary to do a few comparisons.

The `reversed` tag is updated when a credit note is created via the
`ReverseTransaction()` stored procedure.

Partially addressed #1336.