fix(i18n): better cash and credit note description
This commit refactors the cash payments and credit note descriptions.
It makes sure that we have human readable text for our reports to aid
our administrators to better track references by their human readable
equivalents.  This is based on feedback from the administrative team at
Vanga as well as a need to have translated description.

It also slightly improves the cash thermal receipts by ordering the paid
invoices in the order in which they are paid.  The invoices not fully
paid appear at the bottom of the list as one might expect.

Finally, the cash description is now on the receipt.  Icons have been
removed in favor of text.

Partially addresses #1330.