fix(filters) cash registry show daily records
This commit updates the cash registry to include the hot fix patch
introduced in #1339. Only the invoice registry was limited to show daily
records however this is an issue with all commonly used registries.

This commit introduces temporary hot fix code that should be superseeded
by #1347.

This branch should be pulled and manually tested before merging with
master.