refactor(search) Add name search, minimal repsonse
This commit introduces the patients/search/name API. This can be used to
search for a patient by name. It only returns the very minimal data
required to represent a patient, UUID, display name and reference.

This has been applied to the find patient directive when find by name is
selected, drastically reducing the complexity of the query to find a
patient by name.