fix(cash): use semicolons to enable CSV export
This commit fixes a bug in the cash payment registry CSV export caused
by a description that used commas to join lists of invoices together.
By updating to semicolons, this has been fixed.  However, we may wish to
allow a user to change the delimiter of a CSV export file in case as
user-made description breaks the export.