refactor(patients) remove subquery from search
This commit updates the SQL used to search for patients, it also updates
the structure of the table to INDEX and FULLTEXT INDEX columns that are
frequently searched on. This update greatly reduces the number of rows
that must be accessed by the search query, reducing the time required
and working to improve the performance on lower end hardware.