refactor(fs): move to modular fs
This commit updates a few of the client-side modules to make their file
structures more modular.  The following rules have been applied:
 1. Any component that belongs to a module has been moved there.  For
 example, `bhVisit` has been moved to `patients/record`.
 2. Rename components to be more meaningful in context.  For example,
 `bhVisit` became `bhPatientVisit`.
 3. Use John Papa's recommendations on file syntax - snake case is
 preferred over both underscores and capitalization.
 4. Rename files according to function (again from John Papa):
  a. `file.module.js`
  b. `file.routes.js`
  c. `file.component.js`
  d. `file.ctrl.js`
  e. `file.html`

The most changed portions of the application are creditor groups and
patients.  Creditor Groups are now `/creditors/groups` and the patients
gained a whole swath of new components.

Partially addresses #1446.