refactor(trial_balance): use db txns
This commit refactors the trial balance to be more performant by
leveraging the record_uuid for all checks (in binary) and using a
temporary table to JOIN the posting journal instead of expensive IN ()
queries.  This allows the posting_journal to be JOINed using an INDEX.

It also clarifies the logs considerably.  Closes #1428.