fix(grid): enable inline date filtering
The inline date filters were all broken previously during an API change
that did not update the grids with the new API.  Additionally, the old
method did not work due to a syntax issue with the `moment().format()`
call.  This has been corrected by using an upper-cased version of the
bhConstant's date format.

Additionally, this commit bumps ui-grid's version to 4.0.4 to take
advantage of performance improvements.