perf(enterprises): only $digest() on update
This commit implements a first round of performance improvements to the
enterprise module.  It was chosen because it is moderately complex (has
two custom components on it) and little used, ensuring the impact of
bugs are small.

The performance improvement centers around using `ngModelOptions` to
only update inputs on blur.  The child components have been similarly
updated to ignore this - only the non-custom component inputs are
affected.

The initial performance evaluations is:
 - browser.reload(): unchanged (40 - 42 $digests per reload)
 - form.submit(): unchanged (7-8 $digests per submit)
 - form editing: >50% reduction in $digest calls (from 134 to 62
 $digests)

This needs further testing, but it shows that, with relatively small
changes, large performance boosts are possible.

Addresses #1284.