feat(registries): link receipts from references
This commit implements the linked receipts from the grid reference
itself on both cash and invoices. This should save users a click.  Now
cash, patients, invoices, and vouchers are all uniform in their receipt
rendering.

It also makes sure that UUIDs are never rendered in their entirety to
the user.

Finally, the patient links are removed since they are available in the
"actions" dropdown.