feat(filters) Assigning and caching filters
This commit adds an option to toggle off auto parsing statements that do
not have a definition in the server Filter Parser, this allows
additional meta data to be sent through the filters that applies to all
filter (client timestamp).

It also implements modal controller assignement and caching of filters
in the modules service.