feat(journal): the bhAccountSelect is clear-able (#1510)
The bhAccountSelect did not previously support transclusion, leading to a poor experience for users
who wanted to search on the account field.  They would have to clear the entire field (from outside
the search modal) to be able to erase the account select.  This commit allows the account select to
be cleared.