fix(journal): edit transaction doesn't load accounts
This editTransaction() method called for accounts to load even when they did not exist.  This was a
hold over from the old way of editing accounts, which have now been moved to a directive.
Unfortunately, it caused errors to be thrown in the console  and blocked some digests.  This commit
removes the offending code.