feat(journal): default limit to today's records
This commit extends the hacks done for other registries to the Posting Journal, so that it defaults
to today's records as a search.  This should speed up the initial startup of the module considerably
while we develop a better representation of our filters on the client.

Addresses #1436 in a hacky way.