feat(server): verbose MySQL errors (#1521)
Previous versions of BHIMA avoided showing errors to the client,
resulting in a large red bar with no context.  This commit makes sure
that we can at least echo the errors, as well as making sure the client
sees something .. the raw MySQL error.  This will help super users
inform us what kind of errors they are seeing in the meantime.

Additionally, it groups all tests under a new `test.sh` script so that
the "Ctrl-C" interrupt works on unix-based systems to halt the tests.