refactor(billing-services): use account select (#1528)
This commit migrates the Billing Services module to using the account
select component.  By introducing the component, we are able to:
 1. Disabled Title Accounts without any extra effort
 2. Order accounts properly in their nested tree
 3. Remove the AccountService import from both the Create and Update
 controllers of the modal.

In truth, both the create and update state should be combined, but this
is not the commit for it.  This commit paves the way to make the
combination super easy.

Addresses #914 - somehow this module got through the cracks.