refactor(util): remove unused util.js fns
This commit removes unused util.js function from the server library.
Many of these were never used; however, to fully address some of the
changes required, I've extended the `FilterParser` library with a
special function `setGroup()` that will encapsulate the GROUP BY clause
of the SQL.  It turns out pretty cleanly.

I've also refactored the `FilterParser` class to be in conformance with
ESLint.

Closes #1530.  Closes #1531.