feat(vouchers): implement manual Prise En Charge form
This commit implements a manual Prise En Charge form on the Complex Voucher page.  The manual form does not rely on hardware compatibilities of a barcode scanner.  The user must specify an account and a patient, the form will load all the patient's bills, and then correctly assign the patient to the voucher.