fix(journal): do not close search on external click
This commit fixes a usability issue that would automatically close the Journal Search modal when a
user clicked outside of it.  Users often clicked outside the search modal to try and dismiss a
dropdown, but ended up dismissing the entire modal.  Operating on the princple of least surprise
(and least frustration), the dismiss-on-external-click has been removed.

Closes #1617.