feat(account): report opening balances
This commit builds a new feature for accounts that can discern their opening balance (from the
General Ledger) given a date and account id.  The major changes have been added in an "AccountExtra"
file that should be merged into accounts/index.js as soon as the major refactoring is finished.

This PR consists of two parts: the updated account report that now
contains an opening balance and the utility to find the opening balance
give a particular date and account_id.

Closes #1611.  Closes #1636.