refactor(fiscal): best practices for closing
This commit updates the Fiscal Year module with the current best coding
practices.  It specifically targets the closing fiscal year code and
updates it with the following changes:
 1. The translate directive is used instead of the translate filter
 2. The bhAccountSelect is used instead of a custom ui-select
 3. The modal uses a headercrumb(tm)
 4. Removed bhFormDefaults from the form.

This commit addresses most of the points in #1646 for closing a fiscal
year.  Tests have been updated accordingly. The rest of the fiscal year
modules should be updated in a similar fashion.