test(cash): rewrite cash registry tests
This commit re-writes the cash registry tests so that they use the new
filters, search modal.  Two new test harnesses had to be created:
 1. The `FilterService` and its display
 2. The `SearchModal` and its display

The current test syntax uses a giant SearchModal class that contains
some custom code for the Cash Search Modal.  In the future, it might be
prudent to use a mechanism that extends a base SearchModal class.

The FilterServer on the client is now ready for prime-time and can be
leveraged across all registries.

It also would be best to remove the `cash/search` route (and all similar
routes) to make sure that we only have a single GET route.

Partially addresses #1628.