chore(lint): fix server lint issues (#1671)
This commit makes the server pass ESLint. For the most part, all of the
logic remains unchanged; however, the following files have been
completely removed until they can be designed:
 1. `PluginManager` - a hold-out from 1.x.
 2. `errors/index.js` - never actually used on the server.
 3. `finance/ledger.js` - never actually used.

In a few places, ES Lint rules have been added to the top of files to
prevent errors from being thrown for that file.  Ideally, this should be
on a line level.

Closes #1429.