feat(grids): manually saving and restoring state
This commit introduces a new feature to manually save and restore a
grid's state using `ui-grid-save-state`. The feature is prototyped in
the Journal module and adds an entry to the utils dropdown.

State is added as an additional Grid service that can be imported and
used by any module that uses a UI-Grid.