refactor(grids): Remove implicit grid save state
This commit removes the implicit save state from the Journal grid,
called when the columns are changed and groups are updated. This
improves the UX by ensuring the grid settings will only ever be cached
if the user has selected to save the grid state.

It also updates the grouping utility to check if the grid is in a
grouped state. This should be optimised in the future.

Contributes to https://github.com/IMA-WorldHealth/bhima-2.X/issues/1738