refactor(invoices): use FilterService on client
This commit migrates the client side of the Invoice Registry to use the
FilterService.  To do this, the logic is heavily changed in the Invoice
Registry controller, removing the old DeprecatedFilterService and
replacing it completely.  The InvoiceService now manages caching
filters.  Additionally, the form is separated in two tabs with default
filters and specialized filters.

I've also taken the liberty to remove some style hacks and implemented a
search-modal class in bhima-bootstrap.less.  This class simply
transplants the styles to a location where they can be shared.

Progress towards #1749.