fix(reports): Acc. Statement balance and dates
This commit demonstrates one way of tackling #1793. Dates are converted
to the YYYY-MM-DD string before being sent to the server. This ensures
that no conversion of timezones happens between the client and the
server.

This commit also resolves an issue with period totals date calculation
and displays the balance of an account at the end of the statement
report instead of showing individual debits and credits.