fix(reports): Use correct end period comparison
This commit ensures that values the current period are not included in
the period total calculation for determining an accounts opening
balance. This is essential as transactions from the current period (up
to the current date) are then summed and added to the total value. If
the last date in a period is selected this would then result in counting
the current period twice.