feat(reports): Warn on multiple fiscal year span
This commit adds a verification and a warning on account statements that
span multiple fiscal years and target income/expense accounts. Given the
begining balance for every fiscal year will be 0, the cumulative
transaction balance will not line up with the end balance of the
statement. It also resolves issues with dates being sent incorrectly
with the saving method.