test(reports): Add initial files for report checks
This commit adds temporary infrastructure for verifying the data in
reports. These report tests will be run in parrallel with the current
test infrastructure until we have fully introduced running integration
and end to end tests in parrallel and correctly sharding and rebuilding
the database.

This commit also provides an example with Account Statement report
tests.