test: work around timezone corner case
This commit works around a timezone bug that makes testing on other
timezones error out on the first of January.  This is part of a larger
problem of managing timezones and is reported in the issue tracker.  For
now, the work around is to simply change the invoice creation date so
that new contributors can successfully build and test bhima locally.