refactor(db): Add stored proc. for Trial Balance
This commit adds stored procedures for the Trial Balance to the
procedures.sql file.  They all rely in preprocessing transactions to
store the record_uuids in a temporary table.  This should give a
performance improvement as well as increased clarity to the code
(separation of JS concerns versus SQL concerns).