refactor(Trial Balance): combine TB APIs into one
BREAKING CHANGE The Trial Balance APIs have now been combined into a
single route.

By combining routes, we simplify the number of HTTP requests sent for the
Trial Balance information, and allows the client to maintain stateful
information about the errors as well as the summary information.  It
further simplifies the parsing of information sent from the client -
everything is contained in a temporary table shared in the database
transaction.