refactor(journal) remove legacy in-line editing code
This commit refactors the journal and core grid services to rely
on the new feature modal editing and removes legacy structur required
to edit in-line.

It also updates the optimistic updates on the Journal module with a
visual warning for not being able to not having enough information to
update the client (with a prompt to refresh the data set as it may be
out of date).