refactor(bhDateEditor): use onChange() callback
BREAKING CHANGE

This commit migrates the bhDateEditor component to an onChange() style
of callbacks.  This is the recommended Angular 2-style syntax, and
allows the component bindings to dictate the flow of data from `<` and
`@` bindings into the component to `&` for returning data via callbacks.

Additionally, the bhDateEditor has been modernized with `translate`
directives instead of the filter, and everything has been moved into the
$onInit method.

Closes #1960.