refactor(Trial Balance): rewrite Trial Balance
This commit completely rewrites the Trail Balance removing the old
multi-controller headers and replacing it with a single modal controller
and substate controllers for the modal body.  It also introduces the
Trial Balance Service that interacts with the server to load the Trial
Balance.

Key Differences/Features:
 1. Data is loaded a single time - no additional requests are made for
 data-by-account.
 2. Instantaneously switching between error and overview states.
 3. More room on the UI grid display
 4. ui-grid complexity reduce by removing grid grouping and features.
 5. [IN PROGRESS] Links to filter the journal based on the account and
 transaction.

Since all data is loaded on startup, it should remove all issues
reported in #1839.

Closes #1716. Closes #1163. Closes #1839.