fix(links) reference links component updates on change
This commit ensures that the reference link component works in a UI grid
when underlying data is changed. UI Grid optimises the creation of components by
updating existing components instead of replacing them. This requires the
component to act and change the display label on change - it cannot assume
the reference value passed will be static.