refactor(journal) Notify on transaction edit; success and warn
This commit adds success and warn state notifications to transaction editing.
If a transactions was edited a success notification will be displayed and the
logic is adjusted so that if no changes were made a warn notification will be
shown and the selected of the Journal UI Grid will remain unchanged.