fix(links) handle `documentPath` not defined case
This commit adds an additional check to the reference link server
controller; a redirect is not requested unless the document type is
defined and it specifies a documentPath - this stops an infinite
redirect in certain cases.

It also updates the patient document ('PA') to link to the patient
financial history document given a reference lookup.