feat(stock): implement depot selection modal
This commit implements a depot selection modal for the stock modules.
The depot is cached between sessions, but if there is no depot, a user
is immediately directed to the open modal to choose a depot.  This modal
cannot be dismissed if the user has not chosen a depot.

Additionally, a dropdown menu has been implemented on the stock page to
put the "Change Depot" switch in.

Ideally, this should also clear and reset the stock form.  However,
since that isn't even implemented for the "submit" button, this feature
will be added later.

The prototype has been implemented on the Stock Entry page.