fix(bhTransactionTypes): default to empty array
The bhTransactionTypes will now default to an empty array if the
transactions types are not provided to it.  This makes the transaction
types binding truly optional.  A number of small issues have also been
cleaned up: addition of `ngForm` to make sure the `ngModelOptions` is
respected, code updated to standards.