test: end-to-end record deletion
This commit implements end-to-end tests for record deletions.  It also
reworks the 'actions' menus on the cash and invoice registries to use a
similar layout, and builds the 'GridRow' shared class to harness those
menus.  This makes selection of each grid row explicit - the record
human readable reference is passed in to select it.

Closes #2044.