refactor: use bhConstants.defaultFilters
This commit replaces all instances of hand-defined default filters with
the shared instance found in `bhConstants`.  This allows a significant
reduction in code for all services that use the FilterService to limit
records.

Additionally, a few typos were fixed in the affected services, as well
as pruning functions that are never called.  Where appropriate, further
documenting comments were added.

Closes #2133.