fix(vouchers): correctly link reversals
This commit fixes all the voucher links that were previously only
searching on `document_uuid`.  While this is correct for the generic
case of vouchers, for credit notes and debit notes, the reference is not
a direct link.  Instead, the transaction is simply reversed (debits and
credits are swapped) and the reference is stored in the reference_uuid
column.  For this reason, we have to search on both.

Closes #2222.