deps: replace node-uuid with uuid
This commit replaces the deprecated node-uuid with the uuid npm module.
The uuid library is supported through the community, and has much
smaller implementation.  It removes support for `parse()` and
`unparse()` methods, which are not implemented in the `uuid-parse`
library.  Since we use these methods upon occasion ( e.g. `db.bid()`), I
have added that as a dependency as well.

Closes #1395.