refactor(tests): remove duplicate element warnings
This commit addresses as many duplicate element warnings as possible.
There remains warnings on the open debtors report, since the uiSelect
uses an icon to differentiate the sort order, which cannot be selected
upon easily.  Similarly, the complex vouchers cannot be updated easily
with a change to it's cashbox dropdown.

Closes #2193.