feat(Aged Creditors): implement monthly grouping
This commit implements "Group by Month" on the Aged Creditors page.  It
transforms the date range from being a simply 30/60/90 range to being
this month, last month, and the previous month(s).  This allows a much
more intuitive grouping for users who are not used to thinking in ranges
of 30 days.

Closes #1367.