feat(journal): add showFullTransaction filter
The journal filters have been extended with the option to show the
entire transaction following a search on the Journal. This extension
gives the user the ability to not only find an account view, but also a
full transaction view of all the data in the journal.

Note that the implementation is quite heavy (requires two passes on the
Journal).  A warning message has been placed underneath the option to
make the user aware of this.

Closes #2266.