refactor(bhAutoDropup): handle window resize
This commit improves the bhAutoDropup in some corner cases by accurately
estimating the window size on startup and any resize events that occur.
Previously, the first click might have an inaccurate window size.
Furthermore, it implements debounce to ensure that the event loop is not
blocked and the DOM is only manipulated a single time.