feat: implement bhHiddenField
This commit implements the `bhHiddenField` component which hides any
transcluded HTML behind a toggle and lets the user decide when to show
it and to hide it after.  The use case is for option fields in forms
that do not need to be considered by default.  For example, many reports
have complex options that should not be presented on the outset, but can
instead by added in later.

Closes #933.