fix(accounts): improve tests cases
This commit rewrites a few test cases for the Account Management module.
For some reason, the deletion tests did actually work.  Additionally,
the GU.expectGridRows() does not work for grids with large numbers of
rows, as uiGrid will only render visible rows.  This makes testing for
the length of the grid invalid.