test(topic): add unit tests to topic
This commit adds unit tests to the topic library.  It uses a spy plugin
for chai to detect when callbacks have been fired.  In order to wait
for the asynchronous code to complete, each spied function returns a
promise that determines when the testing should proceed.  If no action
happens withing mocha's default timeout, the test fails.

Closes #2314.