fix(patient groups): require notes
This commit fixes the patient groups form making the description a
required field.  It is a required field in the database so this catches
the error on the client before it is even submitted.

Additionally, the client now uses Notify to alert the any errors on data
download or group submission.