feat(reports): additional opts for Account Report
This commit adds a series of additional options for the account reports.
It also improves the code clarity on the server slightly.

The following options are now supported:
 1. Show Original Transaction Currency - This renders the values in the
 currency of the transaction.  Basically, it uses `debit` instead of
 `debit_equiv` and so forth. The opening balances and running totals are
 still in the enterprise currency, however.
 2. Select Total Currency - This allows a user to set a different
 currency to render in the totals footer.  The values are exchanged on
 the date of the report creation.

Together, this allows a very flexible view of the transactions in an
account.

I've also improved the totalling so that the following values are
available:
 1. Opening Balances
 2. Grid Balance (sum of rows)
 3. Grand total (Opening Balances + Grid Balances).

These enhancements strengthen the utility of the account statement and
it is now ready for prime-time.