perf(general_leger): optimize GL Matrix
This commit optimizes the General Ledger matrix query by only doing a
single SELECT with a GROUP BY.  It also removes the need for
pre-querying period ids and uses the raw fiscal year id instead.  This
is the first step to further optimizations of calculating the balances
of title accounts by summing down the account tree.