refactor(fiscal year): close FYs by period totals
This commit implements the closing fiscal year code as a stored
procedure `CloseFiscalYear(fiscalYearId, closingAccountId)`.  This
replaces the previous several hundred lines of JavaScript code with a
single database call, and optimises the process significantly. It also
removes the need for a period number 13 and importing the previous
year's balances.

The following things still need to be done:
 1. Remove the import code for the previous fiscal year balances.
 2. Remove the generation of period number 13.
 3. Document the FY creation procedures with comments.

Closes #1651.  Closes #2464.