fix(journal): cache filters typo
This commit fixes a typo that caused filters to no longer be cached on
the journal.  The code has also been improved by only checking for the
cacheable flag when the filter is attempting to be cached.  This avoids
having to add any values to the modules that use FilterService.

Additionally, I've updated all files I've manipulated to pass ESLint.

Closes #2521.