refactor(fiscal): improve user interface
This commit refactors the fiscal year user interface.  The fiscal year
list is more legible with text and icons as well as proper
capitalization.

The opening balance modal now has totals at the bottom of the grid, bold
title accounts, and the proper currency displayed on the totals.  I've
also refactored the clientside code to be more efficient and brought it
up to ESLint standards.