chore(Tree): optimize internal methods
This commit optimizes the internal tree methods by reducing the
recursion logic to a single function `walk()`.  This allows greater
flexibility - the Tree's API does not need to be changed for every
single new function that needs to operate on the nodes.

The basic API is well documented in the comments.  Please see them and
the unit tests for further usage information.