refactor(fiscal): rewrite opening balance modal
This commit rewrites the opening balance modal using the TreeService
developed on the server for manipulating trees.  The totals for title
accounts and the overall account tree are obtained using these trees.

A potential future optimisation is to only recalculate totals on
subtrees rather than on the entire tree.  This would involve looking up
the parent first, then initialising walk() with the parent node.

For totaling, I currently set all parent balances to 0 before running
totals.  This is inefficient.  We might extend a method on the tree to
allow a preprocessing function to be run before recursing up the tree.

Finally, the opening balances do not include title accounts.

Closes #2536.  Closes #2349.