refactor(credit note): improve modal UI
This commit improves the credit note modal user interface.  It gives
more context about the bill and removes the HTML tables for layout.  It
also attempts to improve the dialogs' text to aid users in making the
right decision.

Finally, the "REVERSAL" text has been changed to "CORRECTION" since that
is both French and English and is easier to refer to.

Closes #767.