refactor(Account Report): support multi currency
This commit implements multi-currency support on the Account Statement
Report. It is derived from a template by @WayneNiles.

The logic is somewhat complex.  If a user produces a report in the
enterprise currency, then the values should convert from a foreign
currency into the enterprise currency.  This is done through _division_
with the formula `amount / rate`.  However, if the user produces a
report in a different currency, all values must be converted from the
enterprise currency to that foreign currency.  This happens by
_multiplication_ with the formula `amount * rate`.

The report has been clarified by:
 1. Combining the two footers into one and naming it `footer`.
 2. Renaming the `openingBalance` property to `header`, since that is
 what is it.
 3. Removing the multiple options on the report page to simply give
 "currency" instead of having different total currencies.
 4. Renaming `accountDetails` to just `account`.

Finally, a new HBS helper has been made for rendering debits/credits
nicely.  It will automatically convert negative values into red currency
values.

Closes #1919.  Closes #2557.