feat(Patients): implement financial activity
This commit implements the patient financial activity bar on the patient
edit page.  The financial activity gives at a glance the amount of time
a patient has been a client of the hospital and their activity over
time.  It also indicates whether the patient has an outstanding debt or
any financial irregularities.