refactor(reports): Cashflow By Service refresh
The Cashflow By Service report is enhanced by updating the client to
ES6 and using the document map to retain the identifiers of cash
payments.  The currency format has also been updated to return a
non-breaking white space, avoiding the tendency to wrap the currency on
FC values.

Partially addresses #2524.