feat(barcode): differentiate between 404/400
This commit differentiates between a "NOT FOUND" barcode (doesn't exist
in the database) and all other errors.  In theory, we shouldn't
encounter a "NOT FOUND" error if users are only scanning barcodes that
are generated by the system, unless someone has deleted it.

This also refactors the tests to remove repetition.