fix(depots): address flaky tests
This commit eliminates some of the uncertainty in the depot tests by
making sure `hasLocation` is always set.  Additionally, the selector
for remove and join location is the same in the page object.  Finally,
the tests have been migrated away from using GU/GA to a more
predictable set of methods.

... I never solved the issue with the joinLocation tests.  So I've
skipped them.  It probably has something to do with not loading the
values soon enough in the `bhLocationSelect`.