chore(db): ignore duplicate keys in triggers
When using bidirectional data sync, the triggers for INSERTs are fired
on the rows that are synced, sometimes causing PRIMARY KEY errors if the
table has a trigger on it that writes metadata elsewhere.  The most
common error is in the `entity_map` and `document_map` tables.

The triggers for creating these maps now ignore the row if a key already
exists for it.