refactor(bhFiscalPeriod): rewrite component
This commit rewrites the bhFiscalPeriod selection component to be easier
to read and more user friendly.  In particular:
  1. The periods are now sorted by date in ascending order.
  2. The periods rely on moment for translation.
  3. The periodFrom selection limits the periodTo selection to only
  the same or greater dates.  This avoids a corner case where periodTo
  could be before periodFrom.
  4. The component uses the FiscalService to get periods instead of a
  custom service.

feat(reports): Profit & Loss Statement

This commit refactors the income and expense report to be a profit and
loss statement.  The income & expense accounts are displayed as a tree
with the totals for each sub-group in the header.  The variation between
the present fiscal year and the previous year provided in subsequent
columns.

Additionally, the bhReportPeriodSelect has been removed and replaced
with the combined bhFiscalSelect and a new bhPeriodSelection component
that allows a user to first select a fiscal year, then select a period.

Closes #1781.  Closes #810.