feat(accounts): implement hidden api
This commit implements and tests the "hidden" column API.  The hidden
column is an extra column to toggle account visibility throughout the
system, but doesn't have a logical (read: accounting) impact on
functioning of the system.

The preliminary work has also been laid out for the dropdown menu in the
account configuration module.

Partially addresses: #2808.