feat(receipts): allow HTML printing
Receipts can now be displayed (and printed, with some conditions) in
HTML.  PDF rendering can take a long time on slower hardware - by
displaying the receipt as HTML, users can reduce the load on the server
and significantly speed up their experience.

The drawback is that receipts printed with HTML are not properly
formatted.  If a user needs a nicely formatted receipt, they are far
better off using PDF for printing.

Closes #2876.