feat(vouchers): add reverse voucher
This commit adds the server side for reversing vouchers. Functionally,
it is identical to reversing the cash payments or invoices, using the
same API.  It also consolidates the grid row templates for all
registries that have reversals into a single template to make things
easier to maintain.

Finally, a bug is fixed with cash reversals that show up with empty
descriptions.

Closes #1204.  Closes #2927.