fix(journal): fixed show full transactions option
This commit changes the value that it is passed into the `onChange`
callback used with the `bh-yes-no` component for the 'Show Full
Transactions' option with the Posting Journal.

Previously the value 0 was always being set which meant the setting
would never reflect a change.

The `default-value` flag was removed from the component as it is
meaningless and not supported, however the current behaviour in
`journal.service.js` is to default `showFullTransactions` to 0 if it has not
yet been defined. A final decision should be made as to the default non-set
value for this flag.