fix(transaction-types): fix type translation issue
This commit fixes a bug on the Transaction Type label in Transaction
Type Management.

Previously the hardcoded `type` value from the database was being
displayed - this was not well formatted and is exclusively in English.
This fix formally defines a constant transaction type map in
`bhConstants` and shares this among controllers that need to dynamically
show types.