perf(transactions): migrate SQL, optimised ID gen
This commit includes database breaking code and bumps the database
version to v0.9.0. It introduces SQL with an optimised core
`GenerateTransactionId` function, this function uses an indexed
MEDIUMINT field to produce an ID in O(1) vs. the O(N) complexity
of the legacy function.

This should have significant performance implications on large datasets.