fix(client-auth): disable partial path matching
This commit fixes an issue with client side authorisation allowing
partial routes that the user does not have permission to, under this
system '/cashboxes' would be a valid path because the user has access to
'/cash'.

This is fixed by adding a second step checking each section of the
required path against the target path, if the base of the target path
match each required section the routing is approved.