feat(cashbox): bh-user component introduced
This commit introduces the `bh-filter` component for displaying an
individual system user.

This is the first component to test a `fetch` architecture that allows
it to either use a details object provided or to optionally take
responsability for fetching that data.

Example usage:
<bh-filter details="user"></bh-filter>
<bh-filter id="5" fetch></bh-filter>

This commit also updates the User service to the latest code layout
standardsd.